arc-overview-tracing: init active_trace only once
It is not necessary to re-set timestamp and window information in the
OnTracingStarted callback. The side effect is that the reported length
of the trace may be a little longer than it otherwise would be. This is
acceptable because we still schedule the trace to stop max-trace-time
seconds after OnTracingStarted is called.

Test: trace on Brya
Test: ArcGraphicsTracingHandlerTest
Bug: None
Change-Id: I36953a740fcfccae80e6b99f5d6aec9c9a536072
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017286
Commit-Queue: Matthew DeVore <matvore@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Alan Ding <alanding@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224517}