desk_profiles: Record the lacros profile ID when saving a desk
RestoreDataCollector is responsible for constructing a saved desk from
the currently active desk. This CL makes it so that the lacros profile
ID that is associated with the current desk is also saved.

Note: this change only handles the lacros profile ID of the desk
itself. Profile IDs for windows *within* the desk will be handled in an
upcoming CL.

BUG=b:308823936

Change-Id: Idc4718db03f69acf80749464284af89fccb1b277
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5021642
Reviewed-by: Yongshun Liu <yongshun@chromium.org>
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Reviewed-by: Judy Wang <shidi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224547}