tote tour: Separate prefs for counterfactual arm
The prefs for the counterfactual arm need to be separated to prevent
overlap. Specifically, if a user were part of the counterfactual arm,
and later the feature was launched to all users, they would never be
shown the nudge of the counterfactual arm used the same prefs as the
enabled arms.

Additionally, this helps create clarity for upcoming metrics that
measure the time between the nudge and the user's use of holding space
features.

Bug: b:274659315
Change-Id: Ib86505ee5e99cf61ddcb18562ecff2e68b71be5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5026277
Commit-Queue: Angus McLean <angusmclean@google.com>
Reviewed-by: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/main@{#1224556}