Improve documentation for Autofill.Funnel.FillAfterSuggestion.*
It was not previously clear in the summary that this metric is recorded
only after the fill happens, not just after the user accepts the
suggestion. These two events can differ, e.g. if the fill requires
authentication and the user either fails or cancels the authentication.

Bug: None
Change-Id: Iba1d4c3d5ea5168cf58a61d4b29c4933a9bd0cf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5028718
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224559}