[uia] GetText for empty button with author specified name
There is a bug in our UIA implementation for TextRangeProvider::GetText
where usually UIA expects embedded object characters for empty
buttons, except if this button has an author specified name.

We were handling this case incorrectly and we were not exposing the
author specified name in empty buttons.

This CL fixes the issue by essentially adding a check that accounts
for this case.

For UIA, we are now exposing the author-specified name for all
empty non-image objects that have an author specified name.

Change-Id: Ic495c380ec430f5a3f853d41c0182e940f6ffe3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4990008
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1224569}