Don't clear Java WebContents before notifying observers of destruction
By destroying the Java WebContents first, a GC happening before the Java
side can clean up its WebContentsObservers can leave things in a
confused state.

We move the clearing of the WebContentsAndroid to after the
WebContentsDestroyed notification, but still before ResetWebContents and
the end of ~WebContentsImpl to still allow the WebContents to be
accessed while invoking the WCA's observers.

Bug: 1366834
Change-Id: I67b3e6a71d218874f124ead9fbda1e289137b096
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5021221
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224574}