blink: Fix crash in HTMLMediaElement.
If a Document is not active, then its ExecutionContext may not be in a
valid state. This CL modifies HTMLMediaElement::ShouldShowControls to
first check if the Document is active.

This CL also adds DCHECKs to HTMLMediaElement and LocalDomWindow.

This CL adds a test. I confirmed that the test reproduces the crash seen
in https://crbug.com/1500204.

Bug: 1485360
Change-Id: I330d0a40a204de11eb04d0125cebdf413155c75e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5019919
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224582}