Always pass DriveFS mount path from ash -> lacros with no /root
GetMountPointPath() and OnMountPointPathChanged() were both passing
the DriveFS 'mount path' to lacros with /root appended, which is
the path for the MyDrive directory, and not the whole of driveFS,
which has other subdirectories like /team_drives.

Other code refers to the /media/fuse/drivefs-<hash> (no /root) as
'mount path' so we should do the same here.

Now ash always passes the DriveFS mount path without /root, and
callers in lacros can add /root if they want to deal with the
MyDrive directory specifically.

There were two ways to get the mount path in lacros, one in the
lacros -> ash direction (GetMountPointPath()), and a newer one in
the ash -> lacros direction (OnMountPointPathChanged()) which
pushes changes to lacros when the path changed in ash.

chrome_paths_lacros.h maintains local state in lacros based on
OnMountPointPathChanged(), which means we always have the latest
path stashed in lacros anyway, meaning there's not much use for
GetMountPointPath() anymore, which is asynchronous, making it
less convenient to use when we need the mount path.

This CL removes uses of GetMountPointPath() and marks it
deprecated. This is also convenient as we don't have to deal with
version skew, as we are only relying on the change in behaviour
for the ash -> lacros direction. All versions of lacros that
will be used with this new behaviour in ash will also have the
new lacros-side code which does not expect the /root dir.

This fixes a bug where the Google Drive MyDrive root folder
would not work when set as the default downloads location in
lacros. It was getting filtered out by
DownloadPrefs::SanitizeDownloadTargetPath(), which was trying
to check if /media/fuse/drivefs-<hash>/root was a child of
itself (rather than a child of just .../drivefs-<hash>).

Bug: 1340762, b/309245269
Change-Id: I4d42918cc50c42df1f542755103b681850eb1509
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5014409
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224595}