Allow pre-warmed tab to be used by any regular tab creation
The initial implementation of pre-warmed tabs required the TabLaunchType to match from warming to use, but we can just set the TabLaunchType
when the tab gets taken and the tab should be usable for any purpose,
including CCT (which will be implemented in future CLs).

This should also lead us to using the pre-warmed tab almost 100% of the time, up from ~75%.

The SpareTab Feature can be used as a kill-switch in the unlikely
event that this breaks things.

Bug: 1496915
Change-Id: I96e4805ea68fdcbf4f1d636660cca26001ed6a78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4995108
Reviewed-by: Sreeja Kamishetty <sreejakshetty@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224608}