[AssistData] Add histograms to measure timeout and node limits
This CL adds two new histograms for the AXTreeSnapshotter. We add
histograms to measure the required max node limit and timeouts for any
given snapshot call to return full data. This will help us to fine-tune
the current limits of max nodes and timeout, which are resulting in
partial data returned to clients. We simply add a second snapshot call
that uses a different callback to record histograms. This is meant to
be low impact, and ideally this histogram will be temporary and then
removed/deprecated.

AX-Relnotes: N/A
Bug: b/303655535
Change-Id: Id10312ae535d0f273f2066f3bfbc3e92f72a730e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5002188
Reviewed-by: David Tseng <dtseng@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224627}