[window-splitting] Fix splitting on extended displays.
There were two issues that caused the bug:

1) When repositioning the windows on drag completion, the bounds were
set using `SetBounds()`, which expected values in root window
coordinates, but the values passed in were in screen coordinates. This
was changed to use `SetBoundsInScreen()`.

2) When dragging a window across different displays, the root window
changes between the call to `Drag()` and `CompleteDrag()`, which changes
the frame of reference of "parent" coordinates. The code was storing the
last drag location in parent coordinates between the two calls, so when
the value was converted back to screen coordinates later, it resulted in
a different screen location. The fix was to store the last drag location
in screen coordinates instead.

Added tests to exercise both conditions above.

Bug: b:306196896
Change-Id: I484defbbcad207817091739f1f8e6e3395475243
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5011489
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Addison Luh <aluh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224679}