reclient_cfgs: fix wrong debug log being collected
With shell=True and passing an array on the first argument to
subprocess.run, the output is for running `cipd` instead of `cipd
auth-info`, which is not really useful for debugging.

Since using shell=True is necessary for Windows cipd.bat, concat the
first argument with space.

Bug: b:304677840
Test: manually
Change-Id: Ie72f47bac64969e9256aecbb5c245cc76e196c75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5029456
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Junji Watanabe <jwata@google.com>
Cr-Commit-Position: refs/heads/main@{#1224717}