Disable ALLOW_MISALIGNED in third_party/unrar
This fixes a host of UBSan errors, such as this one:
https://luci-milo.appspot.com/ui/inv/build-8764440332290274033/test-results?q=FileAnalyzerTest.RarAnalysisResultMetric&sortby=&groupby=

Although the underlying architecture may allow unaligned access, C and
C++ themselves do not allow this. Rather, unaligned loads should be
written with either memcpy, or by endian-agnostic reassembling of values
with shifts and ORs. Modern compilers recognize these patterns and
generate the unaligned load anyway.

There are a couple places where, depending on inlining, the compiler may
not be able to generate the same code. E.g. this function requires the
compiler know that `dest` cannot alias `arg1` or `arg2`.
https://source.chromium.org/chromium/chromium/src/+/main:third_party/unrar/src/rijndael.cpp;drc=b86eab797178b986437ab3db265cceabf0b0496a;l=22

(It would be possible to write that in a way that is both
well-defined and encourages the compiler to load 4 bytes at a time.
Though we'd actually want it to load 16 bytes at a time with SIMD
instructions.)

But I assume unrar is not a hot path for us. If it were, we should
replace that file with calls to BoringSSL anyway, which has actually
optimized implementations. (Rijndael is the original name of AES.)

Bug: 1394755
Change-Id: Id07b3b98f61e282455a1d51c8c45349e074fcb74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5028243
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224718}