[image search] Add search category filter metrics
This cl adds two metrics to record the usage of the search category
filter menu. One of them is the click count on the filter button to
open the menu and the other one is the record the enable state of each
category when the menu is closed and the states commit.

AX-Relnotes: n/a.
Bug: b:308662743
Change-Id: Ifce28d56e91e46bdd443e37ac3180bcadc0aa8a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5009038
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Yulun Wu <yulunwu@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224720}