Re-enable kRestrictCanAccessDataForOriginToUIThread
This CL re-enables Citadel checks in ChildProcessSecurityPolicy.  See
dependent CLs (CL:5005884 and CL:5021611) for prerequisite work that
was required for this.  In particular, the
RequestFileSetCheckedInCanRequestURL kill switch from CL:5005884 is
added as a requirement for the policy function,
ShouldRestrictCanAccessDataForOriginToUIThread(), which determines
whether or not it's ok to assume that CanAccessDataForOrigin() can
only be called on the UI thread, and if not, skip Citadel checks on
the IO thread.  At this point, there is no known code that exercises
CanAccessDataForOrigin() outside of the UI thread and/or relies on
skipped Citadel checks either in production or in tests, though if the
kill switch were activated, this would be needed for the
PrepareDropDataForChildProcess_LocalFiles test.

Bug: 764958, 1286501
Change-Id: Ib77f98c1ad8383362c1532b12edf7b9dc707c007
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5021658
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224727}