Allow creating storage access grants due to the popup heuristic for iframe-initiated popups, under an experiment flag.
We recently discovered that Firefox's implementation enables the popup heuristic for popups loaded by any frame. However, we see potential security concerns with this approach as a 3P iframe can give itself access on the 1P host by simply opening a popup. So we will have this flag-gated behavior on standby in case we need it for compatibility reasons, but we are not currently planning to launch it.

Bug: 1502269
Change-Id: I0902968efecec6cdd293e4e0d59bec8959cb0edc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5026200
Commit-Queue: Anton Maliev <amaliev@chromium.org>
Reviewed-by: Ryan Tarpine <rtarpine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224738}