[//gpu] Pass ClientSI to GPU thread in SIIInProcess::DestroySI(ClientSI)
In preparation for having ClientSharedImage own its associated GMB, this
CL passes the ClientSI instance given to SIIInProcess::DestroySI() to
the GPU thread. This ensures that it will stay alive until the Mailbox
is destroyed, which is when the GMB should be destroyed (and currently
is destroyed).

Bug: 1500611
Change-Id: I45888538ae024f3666d6ab0874349e3e19aca311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5028360
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224797}