Show correct title when updating PSL matched form
Generally Chrome doesn't allow to update PSL matched form and instead saves a copy with a new password value. But when following change
password flow on a website it's possible to update PSL matched form.
This CL ensures that the title is shown correctly in this case.

Bug: 1497918
Change-Id: I8513688053169688308f72fec55ac1f2e8cd7cfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5013569
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1224801}