Get ContentAutofillDriverFactory::DidFinishNavigation early return back.
crrev.com/c/4896011 removed an early return
`ContentAutofillDriverFactory::DidFinishNavigation`.

We believe that this was the root cause of a change in our metrics.
The purpose of this CL is to temporarily bring the lines back to
verify the hypothesis. Once verified, the temporary lines will be
removed again.

We believe that `driver->Reset()` at the end of
`ContentAutofillDriverFactory::DidFinishNavigation` led to resetting
the metric instance variable and that this caused to a change in the
number of submitted samples, which affected metrics of not-submitted
Address and Credit Card forms.

Bug: 1492636
Change-Id: I9e2175f0480f762a418392a94259da1f142722ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5014086
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battre <battre@chromium.org>
Commit-Queue: Dmitry Vykochko <vykochko@google.com>
Cr-Commit-Position: refs/heads/main@{#1224822}