[PEPC] Fix grouped permissions requests and add tests that verify label text
The initial bug was caused by a missing "return" which meant that "equal" variants would still result in the code reaching the end of the function and removing the existing prompt types. Now this behaves correctly.

The rest of the CL includes updates to interactive tests to check the label(s) text for the base flow case as well as separate text-checking tests for the rest of the possible combinations for mic+camera permissions.

Bug: 1462930
Change-Id: I17e3fc4f820f144a0840d4dff851d415d92b72d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5029747
Reviewed-by: Kamila Hasanbega <hkamila@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Thomas Nguyen <tungnh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224832}