[ChromeSigninIntercept] Add icon text during ChromeSigninInterceptBubble
When the Chrome Signin Intercept Bubble is shown, there is a text
displayed next to the avatar icon, such as "Sign in to Chrome?" or
"Sign in to Chromium?".
This was requested as a part of the new sign-in UI.

It looks as follows:
https://screenshot.googleplex.com/Ak6D6Hccth9BW7c.png

This feature is behind the feature flag UnoDesktop.

Bug: b:301386027, 990286
Change-Id: Iae543f7bbc266089d6e1bb6b0588c4c031859a00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994606
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Gabriel Oliveira <gabolvr@google.com>
Commit-Queue: Amelie Schneider <amelies@google.com>
Cr-Commit-Position: refs/heads/main@{#1224852}