Don't include disabled types in the supported types
FormGroup::GetSupportedTypes() return a ServerFieldTypeSet indicating
for which ServerFieldTypes the group can store values. This is used
to determine for which types the FormGroup should vote.

When new types are introduced behind a feature flag, they are generally
only exposed through GetSupportedTypes() when that feature flag is
enabled (example [1]).
For new types added as part of the i18n efforts, this is not the case.
This could potentially be a problem during voting, if a user is enrolled
in an experiment, stores data for such a type and then unenrolls from
the experiment, e.g. by restarting the browser. In this case, that
client would continue sending votes.

This CL fixes this problem by overriding AddressComponent::GetTypes()
in FeatureGuardedAddressComponent, similar to the logic that prevents
storing data in the address component when the feature flag is disabled.

[1] https://source.chromium.org/chromium/chromium/src/+/main:components/autofill/core/browser/data_model/phone_number.cc;l=78;drc=75352e2f484c44c9b4d1f5f9ee827471d4869741

Bug: 1464568
Change-Id: I1e8bfe8e3928b2bc08463f2be2a349e91d3fa568
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017886
Reviewed-by: Dominic Battre <battre@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1224906}