[Passwords] ProcessIncomingSharingInvitationTask takes PasswordForm
Before this CL:
ProcessIncomingSharingInvitationTask processed an incoming password
sharing invitation to decide with it can be approved.

After this CL:
ProcessIncomingSharingInvitationTask processes an incoming password
form instead of an invitation. The invitation is converted to a
password form before passing it over to the task.

This will make follow-up patches that will allow the invitation to
carry multiple passwords easier to implement by spawning a new task
per password form.

This is a mechanical change.

Bug: 1448235
Change-Id: Id5710d5172d2fa3418fe462e811f77b63465a6d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5032395
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224907}