CRD remote command: Make OAuthTokenFetcher reusable
By moving it out of `DeviceCommandStartCrdSessionJob`
it can be used in a future CL to allow reconnecting to an existing CRD
session.

I also changed the logic a little bit so we no longer need to
instantiate a DeviceOAuth2TokenService (and all the classes it depends
on) during unittests.

Bug: b:309446315
Cq-Include-Trybots: luci.chrome.try:linux-chromeos-chrome
Change-Id: Ief21c6901ba5fabff41e50be79f612f4d9c9082d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5018943
Reviewed-by: Edman Anjos <edman@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Cr-Commit-Position: refs/heads/main@{#1224917}